Make build.cmd more like the build.cmd in CoreCLR repo
The CoreFX build.cmd we used as the baseline doesn't need to handle
native build and it can afford to just pass it's command line arguments
to msbuild verbatim. We won't have that luxury in CoreRT since we have
native bits on Windows too.

Fix up project.json files to remove the annoying downgrade warning
messages in Visual Studio.

Change framework in project.json to dotnet, since the other one is
apparently ASP.NET specific.

Disable the framework resource handling that makes our projects include
a phantom Common\src\System\SR.cs file.

I'm not fixing everything with this. We are still using a bunch of
*.target files designed to build the framework to build a compiler,
after all. I don't want to remove that support, because eventually we
will also be building pieces of the framework here, but clearly we need
a way to distinguish between "this is a compiler assembly" and "this is
a framework assembly" at some point.