Add support for running class constructors Similarly to .NET Native, the non-GC static region of a type should be prefixed by a small structure ("static class construction context") that has a pointer to the class constructor and a field indicating whether the class constructor already executed.
This change includes the required JitInterface implementation that decides
whether a class constructor check should be injected by the JIT and a change
in the AsmWriter that emits the extra block and helper call implementation.

I'm sending it for review now, but don't expect to merge it as I want to
wait for David's change. Note that we currently don't set/check the bit indicating
whether the class constructor ran because I don't want to spend time figuring
out how to do it in AT&T assembly syntax.

While this is in review and while David's change is not merged, I'll be looking
into how to hook this up into the managed class constructor runner. CoreLib in
.NET Native has a (commented out) simple version of the runner from the .NET
Native bringup days we could likely successfully use once I implement the
necessary MCG intrinsic.