Add support for Call intrinsic
The concept of Call intrinsic is used in the .NET Native compiler when
C# code needs to perform an indirect call to an address. Since calli is
not exposed in the language, the C# code calls a special method that
will be transformed by the compiler to perform a calli.

The method has shape T Call<T>(IntPtr address, X arg0,... Z argN). The
generated method body loads arguments arg0...argN and performs a calli
to the address given in the first argument.