Add support for InitializeArray intrinsic
Most of the work is piping the RVA static data from the PE image
into our internal data structures, and into the resulting object file.

* Adding IsIntrinsic property that indicates whether the property is an intrinsic
* Adding concept of ReadOnlyDataBlobs to the NodeFactory. I expect this to be more general purpose.
* Name mangling on fields. I was conflicted about adding a big name mangler since we will only need it for RVA statics, but I did it anyway.

This is intentionally only recognizing InitializeArray as an intrinsic. We have some
work to do before we can enable it on e.g. CompareExchange<T>.

We will need to figure out a way to distinguish between JIT intrinsics and
compiler intrinsics in the near future.