Add this pointer to String Ctor methods
Add a 'this' pointer to the static methods that are used as String
constructors to match RyuJIT expectations. RyuJIT passes a useless
'this' in RCX.

This leads to an unfortunate complication in the C++ backend. I tried to
keep the workaround contained in a single place.

Along with the redirection checked in a separate commit, resolves #116.