Half-implementation of RyuJIT generated ROData blocks
Instead of trying to call HandleToObject on something that is not a
handle and throwing an ArgumentOverflowException, capture the ROData
information.

We're still dropping the generated method body on the floor in
Compilation.cs, but the difference is that we no longer throw a managed
exception from something that is called by RyuJIT (which we expect to be
problematic on CoreCLR).

This way we don't need to blacklist 11 methods from getting compiled for
Linux Hello World.