Add more stubs for managed runtime calls
We are compiling more of Hello World, so we need more stubs.

The way we deal with unimplemented functionality in the compiler is that
we abort compilation of the method and emit a method body that has INT 3
in it. More functionality in the compiler means more stuff is getting
compiled, and more of the runtime needs to be present to successfully
link.