Set of fixes that let us compile Hello World on Linux
We need a blacklist of methods that can't be passed to the JIT because
we will throw from JitInterface - on Linux we can't catch the exception
because it involves going across the native stack.

There's also a few methods that don't throw a managed exception, but
crash. The stack was pretty useless. I marked them specially on the
list.

I also added a miserable implementation of System V struct passing
classification. Not sure how correct it is, and it definitely isn't
complete.

You need a protojit.so on Linux to run (yep, remove the "lib" prefix).