Some more Linux Hello World fixes
* Incomplete implementation of write barriers
* Sort-of implementation of platform.linux.cpp. I'm actually not sure if
it works, but might for English.

We still need a fix for #162 to be able to actually use the same
implementation of JitInterface between desktop CLR and CoreCLR.

After that, Hello World on Linux can be compiled, linked, and partially
executed (it will crash on ROData access that we already have a pull
request for).