Add support for constrained method calls
This should be complete for the set of scenarios we can currently
encounter (no shared generics).

* Port of TryResolveConstraintMethodApprox from NUTC/CoreCLR. This is
the part that will need to be extended for shared generics. I think I
managed to come up with an assert that will trigger when the time is
right to update the logic.
* Necessary changes on the JitInterface side
* Implement UncheckedCast in the BCL (I hit this since my fix started
pulling in 30% more methods for hello world).

I'm not checking in a change to allow compiling the extra methods that
we now support (change to _skipJitList in Compilation.cs). This is
because I don't want to spend time updating the list for Linux Hello
World.