Add support for labels in ILEmitter
Not used for anything right now, but I'm working on a piece that would
benefit from this.

We need this for e.g. #89 as well.

I'm not trying to do anything fancy like emitting the short version of
the instructions. We can add that later if we feel like it would be
beneficial. The public surface area shouldn't change as a result of
that.