Improve test run diagnosability
The "throw 42" logic is pretty unfriendly for test run diagnosability -
we can't bucketize the failure. I'm updating it so that we capture some
diagnostic information (the runtime entrypoint called) at the throw
site.

I'm also adding a catch block so that the C++ runtime doesn't pop a
dialog about the unhandled exception and we get a chance to print the
error message.

I'm also fixing a bug in the casting logic that was found in the test
runs. I know the placement of the check is not perf-optimal (or
logical), but hopefully we can call the real casting helpers soon.