Improve usability of ILEmitter
Introduce ILLocalVariable and ILToken types and use them in lieu of int.
This makes the "where do I get the int from" question self-documenting.

It also avoids bugs where, uh... "a friend of mine" uses ldloc instead
of ldarg to load an argument because they didn't check what IntelliSense
autocompleted.

I used an enum instead of a struct with a single integer field because
it's easier to deal with.