Add support for multi-assembly tests
There are a few tests that have helper assemblies that we need to pass
as a reference to ILC. This change will look for those and pass them to
dotnet-compile-native.

The /reference argument is already supported in the wrapper in the CLI
repo, but since we're not using that yet, I added the support to our
dotnet-compile-native.bat.