Fix Unix runtime so that it can be linked to again.
This change fixes issues introduced recently when functions like
ClrVirtualAlloc etc. were added to the gcrhenv.cpp so that the
gcenv.windows.cpp is only used for the sample GC app and not shared
with the runtime.
For Unix though, the gcenv.unix.cpp remained being compiled as part of
the runtime and resulted in multiply defined symbols while linking.
There were also some function signature changes that were not projected
tothe PalRedhawkUnix.cpp. Finally, after removing the gcenv.unix.cpp,
several functions have to be copied from it to the PalRedhawkUnix.cpp.