Real fix for assert !IsStateSet(TSF_Detached)
When I was making the Runtime build on Unix, I have fixed a memset in Thread::Destroy
that was intended to destroy the content of the object in debug builds, but instead
of filling sizeof(*this) bytes, it was filling sizeof(this). Clang issued a warning
at that spot.
But later on, when we have started to actually execute code, it turned out that the
object contents (well, at least its state flags) needs to be preserved since
the DllThreadDetach verifies that the thread was detached using that flag and
asserts since it fails.
@jkotas has commented out the assert as a temporary workaround. Now I am rectifying
it by actually removing the memset and re-enabling the assert.