Fix threshold for PInvoke path in Buffer.Memmove
Threshold for PInvoke path in Buffer.Memmove was way too low. I suspect it was tuned based on the debug builds. I have updated it to be
at the lower end of the break-even point range, based on measurements on several different types of hardware. The best improvements are
on i7-3520M: 64-byte aligned memory blocks - 35% faster, 128 byte aligned memory blocks - 17% faster.

Also:
- Delete misleading comment about PInvoke overhead
- Add comment related to #430