Fix a few small issues
GC code for RO segments had some bugs because it had never been
tested.  CoreCLR doesn't use RO segments.  The fix is to update
sorted_table to keep both the segment memory and the segment
descriptors separately as key/value pairs.  This isn't needed
for normal segments because they are the same, but with RO
segments the memory is in an executable file, while the descriptor
is dynamically allocated.  I also changed "insure" to "ensure".

Changed HndEnumHandles to use a bool parameter instead of BOOL.
This side-steps some differences in the definition of BOOL that
led to some compilation issues.

Port LABELED_RETURN_ADDRESS changes from PN tree.

Fixed a couple of DAC issues.