Add unit test for metadata reader and writer
Whipped up a simple sanity test for metadata readers and writers. This
is mostly a sanity test. We have plenty of runtime coverage for this,
but it's always nice to have a unit test.

This will be useful to give people an idea how to use the metadata
reader and writer APIs since this is a very self-contained sample.