Initial MetadataTransform commit
This transforms the type system object model into the metadata writer
object model. The metadata writer object model can be directly
serialized to a binary blob by the MetadataWriter and used by reflection
at runtime.

Lots of things are not implemented. There are either TODOs or throws in
appropriate places.

I just want to get a few eyes on this before I continue.