Fix two EH bugs
The first issue here is another FailFast problem. Essentially, my last change to RhpPInvokeExceptionGuard--where I tried to account for x86 SEH being different from arm/amd64--failed to handle both modes in which it gets invoked properly. In both cases, we are handling the situation where we've experienced an SEH exception of some kind that hits our backstop handler in managed code. The issue occurs only when we didn't find the exception IP to be in managed code. Usually this means that the exception originated form native code, but this is not true if we, for example, have a call indirect through a bad pointer. The bad pointer is the exception IP in this case and if the caller was managed, we'll be in cooperative mode. In this case, it doesn't make sense to use the current thread's "pinvoke callsite return address". So, for the case where we're in cooperative mode, I've put it back to using the dispatcher context ControlPc on amd64/arm and switched to using the EstablisherFrame on x86. The handler address in the exception registration record is a managed code stub address, so I also had to adjust FindModuleRespectingReturnAddressHijacks to consider all addresses when attempting to find the managed module.

The second issue is a long-standing problem with filter invocation. Preserved registers should not be seeded to the filter and they should most definitely not be restored back to the REGDISPLAY when the filter is done. This is because filter invocation happens during the first pass and, as such, the stackwalker resets its state when it passes over the filter invoke. This means that the REGDISPLAY at the throw site is the 'live' location of any GC references in preserved registers -- i.e. the REGDISPLAY gets updated during GC (not the new stack locations downstream of the filter). So having the filter invocation code trample on the REGDISPLAY values when it's done has the effect of putting back stale values, thus creating a GC hole.

[tfs-changeset: 1561874]