Add GS cookie offset to GC info format.
The old format is binary compatible with the new one, which means programs compiled with the old tool chain should work with the new runtime.  Since we had no free bits left in the first 4 bytes of the header (at least on x64), I repurposed the dynamicAlign bit to indicate there is extra data about the frame.  At present that extra data consist of dynamic stack alignment and GS cookie offset (both of which are optional).  I reserved three bits to allow future extensions of this extra data section.  The gsCookieOffset field is inserted after the epilogCount field to improve memory layout.

CR:  petersol, jkotas

[tfs-changeset: 1566536]