Fix issue that attempting to cast an array to a generic interface with more than one generic parameter fails with a runtime error instead of a correct InvalidCastException.
The problem is that we attempt to find the GenericInstanceDesc of the generic interface type for a non-variant interface that has more than 1 parameter. In that case our toolchain does not provide a GenericInstanceDesc which makes it invalid to call RhGetGenericInstantiation. The fix is to always check the HasGenericVariance flag, as its always set in any case where we may be able to match by variance (as we set the HasGenericVariance flag on all generic interfaces with one generic parameter).

[tfs-changeset: 1568930]