Fir for bug 154284.
The problem is initially caused by the fact that there is no master table that contains a list of all the statically compiled array types. This is ok generally speaking, and the TypeLoader is allowed to build an array EEType dynamically even if it already existed before, because in the runtime, array comparisons/equality checks are based on the equality of the array's element type. (There is a cache in the TypeBuilder to avoid dynamically creating an already existing dynamically allocated array EEType, for perf reasons, not functional reasons).

When arrays start to get used as generic type arguments, this introduces a problem to the TypeBuilder, if it cannot fully resolve the runtime type handle of arrays.
The TypeBuilder can resolve the RTTH of an array type in 2 scenarios:
         a) The TypeSystemContext in use has previously resolved an array RTTH -> type system respresentation (so the TypeDesc representing the array type will have a valid RTTH associated with it)
         b) The array has been dynamically created before by the TypeBuilder, so its RTTH is stored in the TypeBuilder's cache.
The bug happens with generic type instantiations where arrays are used as type arguments when:
         1) The TypeBuilder could not resolve the array's RTTH (none of the 2 scenarios above)
         2) The TypeBuilder has already created the generic type instantiation dynamically previously, when it was able to successfully resolve the array's RTTH to a statically compiled EEType (one of the 2 scenarios above).
This caused the TypeBuilder to re-allocate an existing EEtype for a generic instantiation, and we crashed later during type registration (key already added to dictionary exception).

The same conceptual bug exists with dynamically created method dictionaries.

The fix is to allow the TypeBuilder to search for existing type/method instantiations when array types with no RTTH are used as generic type arguments. This is achieved by loading existing type/method instantiations into the TypeSystem object model, and comparing these existing entries with the input type/method (using TypeSystem object model component comparison instead of typehandle equality checks).

The fix to this bug revealed a weakness in the TypeSystem (debug asser), where in some cases we change the value of the function pointer stored in a MethodDesc (normal canonical <-> universal canonical). This behavior could cause problems (especially in the generic dictionary "Method" cell).
I fixed this by allowing a MethodDesc to have 2 function pointers: one USG and one non-USG. This allows us now to use non-USG method targets for "Method" dictionary cells if we have previously succeeded in loading such a function pointer).

Fixing the USG function pointer issue revealed yet another bug in the binder, where we were not correctly emitting the "FunctionPointerIsUSG" flag for USG entries (bug with unboxing stub entries). Fixed.

[tfs-changeset: 1570044]