Fixing an issue where MissingMetadataException is thrown during databinding with a struct.
We are throwing MissingMetadataException because we are attempting to reflect on IReference<Thickness>. However we should have unboxed\unwrapped this and be reflecting on just a boxed Thickness. We fail to unbox because when the RCW is created by ICLRServices.CreateManagedReference(IReferenceTrackerHost.GetTrackerTarget) we specify the type in the signature is IJupiterObject and we only unbox IUnknown and IInspectable objects.

The signature is actually IUnknown and Jupiter is definitely not passing us a IJupiterObject pointer so I am just updating CreateManagedReference to use the correct signature.

[tfs-changeset: 1571866]