VS debugger testing for feb release found  couple of issues. -Supress throw from GetLastWin32Error -Add some missing functions to external interop -Add IsComObject API
Also this fix  a deadlock in StartupCodeTrigger.Initialize because of the  cyclic dependency between app.dll and app.interop.dll. In normal case this's fine since the runtime will  break the cycle , but if  since we are doing a blocking operation in Initialize  this leads to deadlock. The sequence is , a type from app.dll is being loaded which trigger it's module cctor  ,, calls in to StartupCodeTrigger.Initialize  . Now app.interop.dll's module cctor also calls StartupCodeTrigger.Initialize which leads to the deadlock. Fix is to remove the module cctor from app.interop since all app dll have module cctor which calls in to StartupCodeTrigger.Initialize. Note that just adding an assembly reference to app.interop.dllwont work since module cctor accoring to ECMA is guranteed to be executed only before or at when a type ,filed or function is used , not when's loaded. (it's not DllMain).

[tfs-changeset: 1572488]