Add GetKnown[Method|Field|Type] extension methods
The point is to have a single path through which the compiler locates
things in the core library(ies). This way we can have a different
failure mode when pieces the compiler expects are missing (CoreRT's
fault) vs. pieces the code being compiled expects are missing (user
error, usually).

Once we figure out the exception throwing story, the exception thrown
will be different so that we can handle/translate it properly.