Add reservation API to ObjectDataBuilder
There are situations where ObjectDataBuilder callers might not know the
exact value to emit until later in the object data building process. The
current API surface requires them to precompute the value, which might
potentially be expensive. The new APIs allow to reserve space in the
object data blob and fill it with a value later.

I plan to use this for reflection blobs.