Changed ConditionalWeakTable to use a bump allocator
The code originally used a free list store in the table's next
pointers.  However, due to the lock-free reads, the free list
can never be added to. It is created freshly when copying/resizing.
At this point the entries array can be consolidated, and an index
can be kept for where to insert the next entry. This index can be
incremented until the table is full.

Fix #872