Fix handling of nested TypeReference records
If we have type X that is nested under type Y and:
- Y generates full metadata, and
- X doesn't generate metadata, but we need a TypeReference to it (e.g.
to support DeveloperExperience metadata for MissingMetadataExceptions)

The transform was attempting to set the TypeDefinition as the
ParentNamespaceOrType of the reference. This is not allowed as per the
metadata format. We need to generate a TypeReference for the owning
type.