Avoid livelock in debug builds of MRT
Update the enter_spin_lock_noinstru implementation (which is used to
implement enter_spin_lock in chk/dbg builds) to include the same code
that allows the retail implementation to avoid livelock when a large
number of threads are blocked on the gc_heap::gc_lock.

This change is scoped to chk/dbg builds to avoid perturbing retail
usage of enter_spin_lock_noinstru (which is unreachable and therefore
untestable in the context of MRT).

[tfs-changeset: 1578104]