Clean up the ICompilationRootProvider interface
AddMainMethodCompilationRoot method was kind of odd. It feels like a
single task is getting artificially split into two places. I'm putting
everything into compilation group.

The only reason why StartupCodeMain can't be an implementation detail of
the compilation group is because the CppWriter depends on it. We will
likely be able to make StartupCodeMain fully private to the compilation
group once CppCodegen can deal with export names and refer to it as
__managed_Main.