Complete the timed wait for event
The timed waiting implementation was not finished for OSX, where
we need to use relative time wait for the cond variable to achieve
reliability even when used fiddles with the system time. Also,
autoreset events were not implemented at all.
There was also an arithmetic overflow issue in the TimeSpecAdd.

This change fixes all of these and also ifdefs out the
RhCompatibleReentrantWaitAny for Unix.