Support for compiling to and linking multiple object files
Replace textual section names with ObjectNodeSection objects to make redirecting shared objects to Comdat sections easier.

Support producing full VTables for types. This is used for types shared between modules such as generic instantiations. It is also used for all types emitted in library modules so consuming modules will always have the same layout.

Adjust Optional EEType fields object symbol naming to use an x when a field is not present instead of a 0 since the two cases have different semantics.

In NonGcStaticNode, NonGcStaticNode as the symbol for both it and the CCTor (due to COMDAT limitations that allow only one symbol per section) but apply an offset when a CCtor is present and we need to refer to the nonGc region.

Alter GcStaticsNode to be its own ObjectNode instead of an EmbeddedObjectNode. This is required to support types that go in the shared section since their GcStatics will live there too. Prevent statics from being initialized multiple times. Set the lowest bit in GCStaticDesc's pointers to GCStaticEETypes. At runtime we only initialize GCStatics whose lowest bit is set, masking it off before passing to RhAllocate.  After init is done, the GCStaticDesc's pointer is overwritten with the properly set up GC Static info.  This mechanism allows GCStatics referred to by multiple contributing modules (ie, a common generic instantiation) to only have its statics initialized once.

Replace static VTable helpers with a Node representation of the VTable
for a slice of a given type which chains the base type hierarchy.
Entries are still added on demand as virtual method calls are discovered
in the graph.  However, full VTables are generated when the type is
shared or part of library code. Remove VTable lookup dictionary from NodeFactory

Add an abstract method to ObjectNode: ShouldShareNodeAcrossModules. The idea is that types / methods that can be shared (generics) return True if multi-file compilation is enabled. Other data structures such as constant data can be shared for a potential size on disk win (not measured yet).

In NodeFactory, complete the work to return ExternSymbols when a symbol that lives in another module is requested. Also added ExternMethodSymbolNode for method symbol references cross module.

Stub out various internal runtime implemented methods reachable in System.Private.CoreLib when compiling as a library.