Replace numeric thread id by an opaque one
This change replaces usages of the uint thread id by EEThreadId at
places where we use it for thread matching. It also adds new
PAL function GetPalThreadIdForLogging that returns a numeric thread
id for logging purposes if the target platform supports it.
The reason for this change is that in pthreads, thread id is an
opaque struct and numeric id existence is platform specific.