Fix race condition when running tests
Fixes #974.

Changing the OutputPath also changed ProductPackageDir to something
weird, so I took this as an opportunity to stop placing packaging
related stuff into weird directories (".nuget" was a weird one too -
prefixing with a dot is how you hide directories on Unix).

runtest.sh was passing a dead parameter to restore.sh, so I fixed that
too. I didn't test on Unix. The CI will...