Overhaul delegate thunk handling
In order to properly support the Project N style delegates, I'm changing
a few things around delegates:

* Creating a delegate is no longer a stateless operation. There will now
be a bunch of state associated with every delegate type, wrapped in a
neat `DelegateInfo`.
* `DelegateInfo` will be responsible for managing the various thunks and
the `GetThunk` override.
* I'm getting rid of the delegate shuffle thunk that hardcodes the
destination and replacing it with a Project N style one that uses
indirection through a `calli`.
* `DelegateCreationInfo` will wrap all the low level details and nodes
necessary to call/load from the ReadyToRun helper.
* Ready to run helper emission for the delegate creation helper is now
agnostic to pretty much all the details.