Support ReadyToRun headers on Unix platforms
Linux
Add a linker script which controls how sections are merged to group together module header indirection cells. This allows the ReadyToRun headers to be collated and iterated at start up. The updated CLI looks for a file named "linkerscript" in the CoreRT SDK folder and passes it to the linker on Unix platforms if so.

Remove work-around for Unix from ModulesSectionNode that was inserting symbols for __modules_a and __modules_z which are now consumed from those defined in main.cpp and linked in from libbootstrapper.a

OSX
On Windows / Linux each compiled object file contains a section, .Modules$I, which has a pointer to the ReadyToRun header describing the module's global data. The linker on those platforms collects all object file .Modules$I sections and sandwiches them between two other sections:.Modules$A and .Modules$Z in which we store sentinel symbols available at runtime to iterate the list of ReadyToRun headers.

On OSX the linker doesn't support this model. Instead, a MachO section with the type S_MOD_INIT_FUNC_POINTERS can be created. In that section, place a function pointer. At binary load time, the OS loader will scan for these sections and call all functions in them. Introduce a jump thunk in the compiler that carries the ReadyToRun header address in it. At runtime, it calls a native helper which collects all headers and passes them to the regular startup code.